requirements.txt
setuptools~=58.0.4
pandas~=1.3.4
requests~=2.26.0
bs4~=4.10.0
beautifulsoup4~=4.10.0
sqlalchemy~=1.4.22
tabula-py~=2.3.0
urllib3~=1.26.7
selenium~=3.141.0
numpy~=1.21.2